refactor: solve space circular dependency issue
